feat : 중복 검사 로직에 대해서만 debounce 처리
